requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
pyasn1==0.6.1
pydantic-core==2.23.4
pydantic==2.9.2
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing-extensions==4.12.2